Throw errors from methods when validation fails (#57)
* Throw errors from methods when validation fails

* Extract error creation/checking

* docs(changeset): Reject promises when method handlers are invalid

* Add comment
