+++ .changeset/kind-radios-grin.md
@@ -1,0 +1,5 @@
+---
+"@feltmaps/js-sdk": minor
+---
+
+Reject promises when method handlers are invalid
